massive attack: mp_msg printf format fixes
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@17367 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpcodecs/vf_remove_logo.c
@@ -132,8 +132,8 @@ __________________________________________________________________________
  */
 typedef struct
 {
-  unsigned short int width;
-  unsigned short int height;
+  unsigned int width;
+  unsigned int height;
 
   unsigned char * pixel;
